+++ docs/status/api-contract-coverage-dashboard.json
@@ -17,13 +17,13 @@
     {
       "method": "GET",
       "path": "/api/system/lifecycle",
-      "source": "src/Meridian/UiServer.cs:362",
+      "source": "src/Meridian/UiServer.cs:369",
       "documented": true
     },
     {
       "method": "POST",
       "path": "/api/system/shutdown",
-      "source": "src/Meridian/UiServer.cs:384",
+      "source": "src/Meridian/UiServer.cs:391",
       "documented": true
     },
     {
@@ -941,37 +941,37 @@
     {
       "method": "POST",
       "path": "/api/strategies/covered-call/runs",
-      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:24",
+      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:25",
       "documented": true
     },
     {
       "method": "GET",
       "path": "/api/strategies/covered-call/runs",
-      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:52",
+      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:53",
       "documented": true
     },
     {
       "method": "GET",
       "path": "/api/strategies/covered-call/runs/{runId}/status",
-      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:70",
+      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:71",
       "documented": true
     },
     {
       "method": "GET",
       "path": "/api/strategies/covered-call/runs/{runId}/result",
-      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:93",
+      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:94",
       "documented": true
     },
     {
       "method": "POST",
       "path": "/api/strategies/covered-call/runs/{runId}/cancel",
-      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:135",
+      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:136",
       "documented": true
     },
     {
       "method": "POST",
       "path": "/api/strategies/covered-call/chain-preview",
-      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:158",
+      "source": "src/Meridian.Ui.Shared/Endpoints/CoveredCallEndpoints.cs:159",
       "documented": true
     },
     {
@@ -1839,24 +1839,6 @@
       "documented": true
     },
     {
-      "method": "POST",
-      "path": "/api/ledger/journal-automation/dividend-intake",
-      "source": "src/Meridian.Ui.Shared/Endpoints/LedgerEndpoints.JournalAutomation.cs:19",
-      "documented": true
-    },
-    {
-      "method": "POST",
-      "path": "/api/ledger/journal-automation/fee-accrual-intake",
-      "source": "src/Meridian.Ui.Shared/Endpoints/LedgerEndpoints.JournalAutomation.cs:61",
-      "documented": true
-    },
-    {
-      "method": "POST",
-      "path": "/api/ledger/journal-automation/period-close-intake",
-      "source": "src/Meridian.Ui.Shared/Endpoints/LedgerEndpoints.JournalAutomation.cs:103",
-      "documented": true
-    },
-    {
       "method": "GET",
       "path": "/api/ledger/books",
       "source": "src/Meridian.Ui.Shared/Endpoints/LedgerEndpoints.cs:28",
@@ -2160,6 +2142,24 @@
       "method": "POST",
       "path": "/api/ledger/journal-entry-workbench/lifecycle-action",
       "source": "src/Meridian.Ui.Shared/Endpoints/LedgerEndpoints.cs:2079",
+      "documented": true
+    },
+    {
+      "method": "POST",
+      "path": "/api/ledger/journal-automation/dividend-intake",
+      "source": "src/Meridian.Ui.Shared/Endpoints/LedgerEndpoints.JournalAutomation.cs:19",
+      "documented": true
+    },
+    {
+      "method": "POST",
+      "path": "/api/ledger/journal-automation/fee-accrual-intake",
+      "source": "src/Meridian.Ui.Shared/Endpoints/LedgerEndpoints.JournalAutomation.cs:61",
+      "documented": true
+    },
+    {
+      "method": "POST",
+      "path": "/api/ledger/journal-automation/period-close-intake",
+      "source": "src/Meridian.Ui.Shared/Endpoints/LedgerEndpoints.JournalAutomation.cs:103",
       "documented": true
     },
     {
@@ -5353,11 +5353,6 @@
       "documented": true
     },
     {
-      "name": "NullReportingRunNotifier",
-      "source": "src/Meridian.Contracts/Workstation/IReportingRunNotifier.cs:15",
-      "documented": true
-    },
-    {
       "name": "InvestmentAccountingTransactionKindDto",
       "source": "src/Meridian.Contracts/Workstation/InvestmentAccountingTransactionLabDtos.cs:6",
       "documented": true
@@ -5395,6 +5390,11 @@
     {
       "name": "InvestmentAccountingTransactionLabPreviewDto",
       "source": "src/Meridian.Contracts/Workstation/InvestmentAccountingTransactionLabDtos.cs:71",
+      "documented": true
+    },
+    {
+      "name": "NullReportingRunNotifier",
+      "source": "src/Meridian.Contracts/Workstation/IReportingRunNotifier.cs:15",
       "documented": true
     },
     {
