+++ docs/status/doc-health-dashboard.json
@@ -1,6 +1,6 @@
 {
   "total_files": 674,
-  "total_lines": 96420,
+  "total_lines": 96469,
   "orphaned_count": 330,
   "no_heading_count": 148,
   "stale_count": 0,
@@ -2405,7 +2405,7 @@
     },
     {
       "path": "docs/generated/repository-structure.md",
-      "line_count": 8663,
+      "line_count": 8705,
       "has_heading": true,
       "todo_count": 9,
       "last_modified_utc": "1970-01-01T00:00:00+00:00",
@@ -5461,7 +5461,7 @@
     },
     {
       "path": "src/Meridian.Execution.Sdk/README.md",
-      "line_count": 78,
+      "line_count": 80,
       "has_heading": true,
       "todo_count": 1,
       "last_modified_utc": "1970-01-01T00:00:00+00:00",
@@ -5469,7 +5469,7 @@
     },
     {
       "path": "src/Meridian.Execution/README.md",
-      "line_count": 96,
+      "line_count": 99,
       "has_heading": true,
       "todo_count": 1,
       "last_modified_utc": "1970-01-01T00:00:00+00:00",
@@ -5613,7 +5613,7 @@
     },
     {
       "path": "src/Meridian.Risk/README.md",
-      "line_count": 63,
+      "line_count": 65,
       "has_heading": true,
       "todo_count": 1,
       "last_modified_utc": "1970-01-01T00:00:00+00:00",
